requirements.txt
annoy==1.17.0
faiss_cpu==1.7.4
matplotlib==3.8.0
numpy==1.22.4
pandas==2.1.3
python_igraph==0.11.3
scipy==1.8.0
umap_learn==0.5.4
